Implement development exception middleware https://github.com/SoftUni-Internal/exam-systems-issues/issues/224
